requirements.txt
numpy==1.21.6
matplotlib
tensorflow-cpu==2.9
openpyxl